+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsAboutFragment.java
@@ -93,6 +93,11 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
         flags.put(strRes("piko_pref_clear_tracking_params"),SettingsStatus.cleartrackingparams);
         flags.put(strRes("piko_pref_unshorten_link"),SettingsStatus.unshortenlink);
         flags.put(strRes("piko_pref_force_translate"),SettingsStatus.forceTranslate);
+        flags.put(strRes("piko_pref_round_off_numbers"),SettingsStatus.roundOffNumbers);
+        flags.put(strRes("piko_pref_customisation_inlinetabs"),SettingsStatus.inlineBarCustomisation);
+        flags.put(strRes("piko_pref_debug_menu"),SettingsStatus.enableDebugMenu);
+        flags.put(strRes("piko_pref_hide_premium_prompt"),SettingsStatus.hidePremiumPrompt);
+        flags.put(strRes("piko_pref_hide_hidden_replies"),SettingsStatus.hideHiddenReplies);
 
         LegacyTwitterPreferenceCategory patPref = preferenceCategory(strRes("piko_pref_patches"), screen);