+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsFragment.java
@@ -190,6 +190,16 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 );
             }
 
+            if (SettingsStatus.hidePremiumPrompt) {
+                adsPrefs.addPreference(
+                        switchPreference(
+                                strRes("piko_pref_hide_premium_prompt"),
+                                "",
+                                Settings.ADS_HIDE_PREMIUM_PROMPT
+                        )
+                );
+            }
+
 
         }
 
@@ -264,6 +274,26 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 );
             }
 
+            if (SettingsStatus.roundOffNumbers) {
+                miscPrefs.addPreference(
+                        switchPreference(
+                                strRes("piko_pref_round_off_numbers"),
+                                strRes("piko_pref_round_off_numbers_desc"),
+                                Settings.MISC_ROUND_OFF_NUMBERS
+                        )
+                );
+            }
+
+            if (SettingsStatus.enableDebugMenu) {
+                miscPrefs.addPreference(
+                        switchPreference(
+                                strRes("piko_pref_debug_menu"),
+                                "",
+                                Settings.MISC_DEBUG_MENU
+                        )
+                );
+            }
+
             if (SettingsStatus.featureFlagsEnabled) {
                 miscPrefs.addPreference(
                         buttonPreference(
@@ -282,7 +312,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 customisationPrefs.addPreference(
                         multiSelectListPreference(
                                 strRes("piko_pref_customisation_profiletabs"),
-                                "",
+                                strRes("piko_pref_import_warn"),
                                 Settings.CUSTOM_PROFILE_TABS
                         )
                 );
@@ -291,7 +321,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 customisationPrefs.addPreference(
                         listPreference(
                                 strRes("piko_pref_customisation_timelinetabs"),
-                                "",
+                                strRes("piko_pref_import_warn"),
                                 Settings.CUSTOM_TIMELINE_TABS
                         )
                 );
@@ -300,7 +330,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 customisationPrefs.addPreference(
                         multiSelectListPreference(
                                 strRes("piko_pref_customisation_sidebartabs"),
-                                "",
+                                strRes("piko_pref_import_warn"),
                                 Settings.CUSTOM_SIDEBAR_TABS
                         )
                 );
@@ -310,8 +340,18 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 customisationPrefs.addPreference(
                         multiSelectListPreference(
                                 strRes("piko_pref_customisation_navbartabs"),
-                                "",
+                                strRes("piko_pref_import_warn"),
                                 Settings.CUSTOM_NAVBAR_TABS
+                        )
+                );
+            }
+
+            if (SettingsStatus.inlineBarCustomisation) {
+                customisationPrefs.addPreference(
+                        multiSelectListPreference(
+                                strRes("piko_pref_customisation_inlinetabs"),
+                                strRes("piko_pref_import_warn"),
+                                Settings.CUSTOM_INLINE_TABS
                         )
                 );
             }
@@ -407,6 +447,15 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 );
             }
 
+            if (SettingsStatus.hideHiddenReplies) {
+                timelinePrefs.addPreference(
+                        switchPreference(
+                                strRes("piko_pref_hide_hidden_replies"),
+                                "",
+                                Settings.TIMELINE_HIDE_HIDDEN_REPLIES
+                        )
+                );
+            }
         }
 
 
@@ -441,7 +490,23 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
                 )
         );
 
-        //export section
+        backupPref.addPreference(
+                buttonPreference(
+                        strRes("delete")+": "+strRes("notification_settings_preferences_category"),
+                        "",
+                        Settings.RESET_PREF.key
+                )
+        );
+
+        backupPref.addPreference(
+                buttonPreference(
+                        strRes("delete")+": "+strRes("piko_title_feature_flags"),
+                        "",
+                        Settings.RESET_FLAGS.key
+                )
+        );
+
+        //about section
         LegacyTwitterPreferenceCategory aboutPref = preferenceCategory(strRes("piko_title_about"), screen);
         aboutPref.addPreference(
                 buttonPreference(
@@ -504,6 +569,9 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         }else if (key == Settings.VID_MEDIA_HANDLE.key) {
             entries = Utils.getResourceStringArray("piko_array_download_media_handle");
             entriesValues = new CharSequence[]{"download_media","copy_media_link", "always_ask"};
+        }else if (key == Settings.CUSTOM_INLINE_TABS.key) {
+            entries = Utils.getResourceStringArray("piko_array_inlinetabs");
+            entriesValues = new CharSequence[]{"Reply","Retweet", "Favorite","ViewCount","AddRemoveBookmarks", "TwitterShare"};
         }
         preference.setEntries(entries);
         preference.setEntryValues(entriesValues);
@@ -529,6 +597,9 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         }else if (key == Settings.CUSTOM_NAVBAR_TABS.key) {
             entries = Utils.getResourceStringArray("piko_array_navbar");
             entriesValues = new CharSequence[]{"HOME","GUIDE", "SPACES","COMMUNITIES","NOTIFICATIONS","CONNECT","COMMUNITY_NOTES","BOOKMARKS","DMS","GROK","MEDIA_TAB"};
+        }else if (key == Settings.CUSTOM_INLINE_TABS.key) {
+            entries = Utils.getResourceStringArray("piko_array_inlinetabs");
+            entriesValues = new CharSequence[]{"Reply","Retweet", "Favorite","ViewCount","AddRemoveBookmarks", "TwitterShare"};
         }
         preference.setEntries(entries);
         preference.setEntryValues(entriesValues);
@@ -550,7 +621,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
             app.revanced.integrations.twitter.Utils.startUndoPostActivity();
         } else if (key.equals(Settings.PREMIUM_ICONS.key)) {
             app.revanced.integrations.twitter.Utils.startAppIconNNavIconActivity();
-        }else if (key.equals(Settings.MISC_FEATURE_FLAGS.key)) {
+        } else if (key.equals(Settings.MISC_FEATURE_FLAGS.key)) {
             startFragment(new FeatureFlagsFragment());
         } else if (key.equals(Settings.EXPORT_PREF.key)) {
             startBackupFragment(new BackupPrefFragment(), false);
@@ -560,8 +631,12 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
             startBackupFragment(new RestorePrefFragment(), false);
         } else if (key.equals(Settings.IMPORT_FLAGS.key)) {
             startBackupFragment(new RestorePrefFragment(), true);
-        }else if (key.equals(Settings.PATCH_INFO.key)) {
+        } else if (key.equals(Settings.PATCH_INFO.key)) {
             startFragment(new SettingsAboutFragment());
+        } else if (key.equals(Settings.RESET_PREF.key)) {
+            app.revanced.integrations.twitter.Utils.deleteSharedPrefAB(context,false);
+        } else if (key.equals(Settings.RESET_FLAGS.key)) {
+            app.revanced.integrations.twitter.Utils.deleteSharedPrefAB(context,true);
         }
 
         return true;